CLASS: FeeTable, TEST: getFeeList SUCCESS - changed fee inside the Fee class from double to int

+++ src/server/logic/model/Fee.java
@@ -2,9 +2,9 @@ package server.logic.model;
 
 public class Fee {
 	int userId;
-	double fee;
+	int fee;
 	
-	public Fee(int userId, double fee) {
+	public Fee(int userId, int fee) {
 		this.setUserId(userId);
 		this.setFee(fee);
 	}
@@ -13,7 +13,7 @@ public class Fee {
 	public void setUserId(int userId) {
 		this.userId = userId;
 	}
-	public void setFee(double fee) {
+	public void setFee(int fee) {
 		this.fee = fee;
 	}
 	//
@@ -21,7 +21,7 @@ public class Fee {
 	public int getUserId() {
 		return userId;
 	}
-	public double getFee() {
+	public int getFee() {
 		return fee;
 	}
 }
